requirements.txt
Pillow==9.4.0
argparse==1.4.0